Migrate Community to useNotificationData
